requirements.txt

apache-airflow==2.8.1
pandas
apache-airflow-providers-amazon
apache-airflow[celery]==2.8.1
